requirements.txt
# Core
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0

# Computer Vision
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
opencv-python>=4.9.0

# Machine Learning
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0

# Explainability
shap>=0.44.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0

# Utilities
tqdm>=4.66.0
